refactor basic app layout
