+++ atramhasis/static/admin/src/app/ui/AppUi.html
@@ -1,0 +1,1 @@
+<div>AppUi</div>
