+++ atramhasis/static/admin/src/app/ui/AppUi.js
@@ -1,0 +1,88 @@
+/**
+ * Main application user interface.
+ */
+define([
+  'dojo/_base/declare',
+  'dojo/_base/lang',
+  'dojo/_base/fx',
+  'dojo/dom-style',
+  'dojo/topic',
+  'dijit/_WidgetBase',
+  'dijit/_TemplatedMixin',
+  'dojo/text!./AppUi.html'
+], function (declare, lang, fx, domStyle, topic, _WidgetBase, _TemplatedMixin, template) {
+  return declare([_WidgetBase, _TemplatedMixin], {
+
+    templateString: template,
+    loadingContainer: null,
+
+    /**
+     * Standard widget function.
+     * @public
+     */
+    postCreate: function () {
+      this.inherited(arguments);
+      console.debug('AppUi::postCreate');
+      this._registerLoadingEvents();
+
+    },
+
+    /**
+     * Standard widget function.
+     * @public
+     */
+    startup: function () {
+      this.inherited(arguments);
+      console.debug('AppUi::startup');
+
+      this._hideLoading();
+    },
+
+    /**
+     * Hide the 'Loading'-overlay.
+     * @public
+     */
+    _hideLoading: function () {
+      var node = this.loadingContainer;
+      fx.fadeOut({
+        node: node,
+        onEnd: function (node) {
+          domStyle.set(node, 'display', 'none');
+        },
+        duration: 1000
+      }).play();
+    },
+
+    /**
+     * Show the 'Loading'-overlay.
+     * @public
+     */
+    _showLoading: function (message) {
+      if (!message) message = "";
+      var node = this.loadingContainer;
+      query(".loadingMessage", node).innerHTML(message);
+
+      domStyle.set(node, 'display', 'block');
+      fx.fadeIn({
+        node: node,
+        duration: 1
+      }).play();
+    },
+
+    /**
+     * Listen to events to show/hide the loading overlay
+     * @private
+     */
+    _registerLoadingEvents: function () {
+      this.own(
+        topic.subscribe('standby.show',lang.hitch(this, function(evt){
+          this._showLoading(evt.message);
+        })),
+        topic.subscribe('standby.stop',lang.hitch(this, function(){
+          this._hideLoading();
+        }))
+      );
+
+    }
+  });
+});
