+++ atramhasis/static/css/app-admin.css
@@ -13229,6 +13229,50 @@ ul.downloadtop {
     word-break: break-word;
   }
 }
+/* main html */
+/* line 2, ../scss/atramhasis/_atramhasis-admin.scss */
+html, body {
+  width: 100%;
+  height: 100%;
+}
+
+/* noscript div */
+/* line 5, ../scss/atramhasis/_atramhasis-admin.scss */
+.noscript {
+  top: 0;
+  left: 0;
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  z-index: 1002;
+  display: block;
+  background-color: #efefef;
+  text-align: center;
+  padding: 25px;
+}
+
+/* Loading overlay */
+/* line 8, ../scss/atramhasis/_atramhasis-admin.scss */
+.pageOverlay {
+  top: 0;
+  left: 0;
+  position: absolute;
+  height: 100%;
+  width: 100%;
+  z-index: 1001;
+  display: block;
+  background-color: #666666;
+  text-align: center;
+  color: #d6d6d6;
+}
+
+/* line 9, ../scss/atramhasis/_atramhasis-admin.scss */
+.loadingMessage {
+  font-size: 1em;
+  margin: 30px;
+  color: #d6d6d6;
+}
+
 /* line 1, ../scss/_atramhasis.scss */
 ul.no_bullet {
   list-style-type: none;
@@ -13270,48 +13314,4 @@ h4.relations-subtitle {
   font-size: 20px;
   margin-bottom: 10px;
   margin-right: 5px;
-}
-
-/* main html */
-/* line 18, ../scss/app-admin.scss */
-html, body {
-  width: 100%;
-  height: 100%;
-}
-
-/* noscript div */
-/* line 21, ../scss/app-admin.scss */
-.noscript {
-  top: 0;
-  left: 0;
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  z-index: 1002;
-  display: block;
-  background-color: #efefef;
-  text-align: center;
-  padding: 25px;
-}
-
-/* Loading overlay */
-/* line 24, ../scss/app-admin.scss */
-.pageOverlay {
-  top: 0;
-  left: 0;
-  position: absolute;
-  height: 100%;
-  width: 100%;
-  z-index: 1001;
-  display: block;
-  background-color: #666666;
-  text-align: center;
-  color: #d6d6d6;
-}
-
-/* line 25, ../scss/app-admin.scss */
-.loadingMessage {
-  font-size: 1em;
-  margin: 30px;
-  color: #d6d6d6;
 }
